requirements.txt
Flask>=2.3.3
pdfplumber>=0.9.0
python-docx>=0.8.11
pypdf>=3.10.0
gunicorn>=21.2.0

